requirements.txt
flask
gunicorn
python-dotenv
requests
Pillow